+++ fix/src/Indexing.mli
@@ -39,6 +39,25 @@ type 'n cardinal
    added. *)
 val cardinal : 'n cardinal -> int
 
+(**[is_fixed n] determines whether the cardinal [n] is (or has been) fixed. *)
+val is_fixed : 'n cardinal -> bool
+
+(**The equality type. *)
+type (_, _) eq = Refl : ('a, 'a) eq
+
+(**[equal n m] tests whether the cardinals [n] and [m] are equal, and if so,
+   produces a type-level equality between the types ['n] and ['m]. If the
+   cardinals [n] or [m] are not yet fixed, then, as a side effect of this
+   call, they become fixed. *)
+val equal : 'n cardinal -> 'm cardinal -> ('n, 'm) eq option
+
+(**[assert_equal n m] checks that the cardinals [n] and [m] are equal, and
+   produces a type-level equality between the types ['n] and ['m]. If the
+   cardinals [n] or [m] are not yet fixed, then, as a side effect of this
+   call, they become fixed.
+   @raise Invalid_argument if the cardinals are not equal. *)
+val assert_equal : 'n cardinal -> 'm cardinal -> ('n, 'm) eq
+
 (**If [n] is a type-level name for a finite set, then a value [i] of type
    [n index] is an integer value that is guaranteed to inhabit the set [n].
 
@@ -132,7 +151,8 @@ module Index : sig
      an integer [i] of type [int] into an index: that is, [of_int n i] returns
      [i] at type [n index]. The integer [i] must lie in the semi-open interval
      [\[0, n)]. This is enforced by a runtime check. Calling [of_int n i]
-     fixes the cardinal [n]. *)
+     fixes the cardinal [n].
+     @raise Invalid_argument if the index [i] is out of bounds. *)
   val of_int : 'n cardinal -> int -> 'n index
 
   (**{!to_int} casts an index [i] back to an ordinary integer value. *)
@@ -141,6 +161,10 @@ module Index : sig
   (**[iter n yield] calls [yield i] successively for every index in the range
      [\[0, n)], in increasing order. *)
   val iter : 'n cardinal -> ('n index -> unit) -> unit
+
+  (**[rev_iter n yield] calls [yield i] successively for every index in the
+     range [\[0, n)], in decreasing order. *)
+  val rev_iter : 'n cardinal -> ('n index -> unit) -> unit
 
   (**This exception is raised by an iterator (created by {!enumerate}) that is
      queried after it has been exhausted. *)
@@ -156,13 +180,16 @@ end
 
 (**A vector of type [(n, a) vector] is a (fixed-size) array whose indices lie
    in the type-level set [n] and whose elements have type [a]. *)
-type ('n, 'a) vector =
-  private 'a array
+type ('n, 'a) vector
 
 (**The submodule {!Vector} allows safely manipulating indices into a vector. *)
 module Vector : sig
 
   type ('n, 'a) t = ('n, 'a) vector
+
+  (**[as_array v] exposes a view of the vector [v] as an ordinary array.
+     This is a safe cast. This operation does nothing at runtime. *)
+  val as_array : (_, 'a) t -> 'a array
 
   (**{!length} is analogous to [Array.length], but returns a cardinal instead
      of an ordinary integer. *)
@@ -196,7 +223,69 @@ module Vector : sig
      cardinal [n]. *)
   val init : 'n cardinal -> ('n index -> 'a) -> ('n, 'a) t
 
-  (**{!map} is analogous to [Array.map]. *)
+  (**{!map} is [Array.map]. *)
   val map : ('a -> 'b) -> ('n, 'a) t -> ('n, 'b) t
+
+  (**{!mapi} is [Array.mapi]. *)
+  val mapi : ('n index -> 'a -> 'b) -> ('n, 'a) t -> ('n, 'b) t
+
+  (**{!copy} is [Array.copy]. *)
+  val copy : ('n, 'a) t -> ('n, 'a) t
+
+  (**{!iter} is [Array.iter]. *)
+  val iter : ('a -> unit) -> ('n, 'a) t -> unit
+
+  (**{!iteri} is [Array.iteri]. *)
+  val iteri : ('n index -> 'a -> unit) -> ('n, 'a) t -> unit
+
+  (**{!iter2} is [Array.iter2]. *)
+  val iter2 : ('a -> 'b -> unit) -> ('n, 'a) t -> ('n, 'b) t -> unit
+
+  (**{!fold_left} is [Array.fold_left]. *)
+  val fold_left : ('a -> 'b -> 'a) -> 'a -> (_, 'b) t -> 'a
+
+  (**{!fold_right} is [Array.fold_right]. *)
+  val fold_right : ('b -> 'a -> 'a) -> (_, 'b) t -> 'a -> 'a
+
+  (**[fold_left2 f accu v1 v2] folds the function [f] with initial accumulator
+     [accu] simultaneously over the vectors [v1] and [v2]. The elements of the
+     vectors are processed left to right. *)
+  val fold_left2 : ('a -> 'b -> 'c -> 'a) -> 'a -> ('n, 'b) t -> ('n, 'c) t -> 'a
+
+  (**[fold_right2 f v1 v2 accu] folds the function [f] with initial
+     accumulator [accu] simultaneously over the vectors [v1] and [v2]. The
+     elements of the vectors are processed right to left. *)
+  val fold_right2 : ('b -> 'c -> 'a -> 'a) -> ('n, 'b) t -> ('n, 'c) t -> 'a -> 'a
+
+  (**{!to_list} is [Array.to_list]. *)
+  val to_list : (_, 'a) t -> 'a list
+
+  (**{!sort} is [Array.sort]. *)
+  val sort : ('a -> 'a -> int) -> ('n, 'a) t -> unit
+
+  (**Suppose [v] is a vector that maps indices in the range [\[0,m)]
+     to indices in the range [\[0,n)], and suppose that [v] represents
+     an injective function. Then, [invert n v] returns a vector that
+     represents the inverse function, a partial function of [\[0,n)]
+     into [\[0,m)]. *)
+  val invert : 'n cardinal -> ('m, 'n index) t -> ('n, 'm index option) t
+
+  (**[of_array n a] checks that the cardinal [n] is equal to the length of the
+     array [a] and converts the array [a] to a vector. If the cardinal [n] is
+     not yet fixed, then, as a side effect of this call, it becomes fixed.
+     @raise Invalid_argument if the cardinal and array length are not equal. *)
+  val of_array : 'n cardinal -> 'a array -> ('n, 'a) t
+
+  (**The module type [V] is a module-level analogue of the type [vector]. *)
+  module type V = sig
+    type n
+    type a
+    val vector : (n, a) t
+  end
+
+  (**The functor [Of_array] converts an ordinary array to a vector. No runtime
+     check is involved. The type component [n] of the result serves as a fresh
+     type-level identifier for the length of this vector. *)
+  module Of_array (A : sig type a val array : a array end) : V with type a = A.a
 
 end